pyproject.toml
[project]
name = "payslip-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "manweljs",email = "80235442+manweljs@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "strawberry-graphql (>=0.273.2,<0.274.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "aioodbc (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
